Начало установки AdminLTE

+++ app/Http/Controllers/TbComplexController.php
@@ -53,6 +53,11 @@ class TbComplexController extends Controller
     public function destroy($id) // Для удаления конкретного ресурса из БД
     {
         $complex = TbComplex::find($id);
+        
+        if (!$complex) {
+            return response()->json(['error'=>'Такого комплекса нету'], 404);
+        }
+
         $complex->delete();
 
         return response()->json(['message' => 'Complex deleted']);
